+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All Notable changes to `AssetLibrary` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## NEXT - 2018-07-24
+
+### Added
+- The addFile and addFiles method on the assetTrait now return the added assets.
+- Filenames will now be slugified to avoid weird behaviour when accessing the files throught urls.
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## NEXT - YYYY-MM-DD
 
 ### Added